Added a class for the inverter chain path.
Moved the monte carlo simulation into it's own function.

Now we have a test script (inversor_chain_path_monte_carlo_sim.py),
which instantiates the path and passes it to the monte carlo sim
function, which runs the sim on that path.

This means the monte carlo simulation is independant of the path and the
tech.