Added gate and path tests.
This allows you to check that your gate / path function as expected.
As part of this I added a second path: a chain of 5 nand gates. It's not
a serious path that I'm suggesting we test (although we could), it was
just something to make sure the path test works as expected.